refactor(core): split shared.tsx into one file per export
Separate plain .ts modules (constants, date formatters, buildTextFooter)
from .tsx component files, and give each react-email component
(Masthead, ReserveButton, SpecialBadge, BrowseCalendarFooter, EmailFooter)
its own file under packages/core/emails/shared/, so each is easy to find
by name instead of scrolling one combined file.

Verified behavior-neutral: rendered both templates with the existing
preview scripts before and after and diffed the HTML/text output --
byte-identical in both cases.
