+++ packages/core/emails/shared/buildTextFooter.ts
@@ -1,0 +1,7 @@
+import { MANAGE_URL } from "./constants";
+
+export function buildTextFooter(reason: string) {
+  return `---
+${reason}
+Manage notification settings: ${MANAGE_URL}`;
+}
